layout 리팩토링
