+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import '../styles/globals.css'
 import React from 'react'
 import blogMetaData from '@/metadata/blogMetaData'
 import UiContextProvider from '@/components/ui-layout/UiContextProvider'
-import { DynamicLayout } from '@/components/ui-layout/DynamicLayout'
+import { DynamicLayout } from '@/components/ui-layout/main/DynamicLayout'
 import { AuthSessionProvider } from '@/components/auth/AuthSessionProvider'
 import SideBarProvider from '@/components/ui-layout/sidebar/SiderBarProvider'
