+++ src/components/memo/folder_navigator/header/NewMemoLink.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React, { type Dispatch, type SetStateAction, useContext, useEffect } from 'react'
-import { TabBarContext } from '@/components/ui-layout/DynamicLayout'
+import { TabBarContext } from '@/components/ui-layout/main/DynamicLayout'
 import { createMemo } from '@/api/memo'
 import { type FolderInfo, type SimpleMemoInfo } from '@/api/models'
 import Image from 'next/image'
